refactor(desktop): extract PromptPreviewModal from App.tsx (F4.2)
Lift the analyzer-prompt preview into its own component, moving the
keyboard-scroll ref + window-listener effect in with it so it's self-contained
(props: text + onClose). App.tsx -41 lines, and one App-level useEffect + ref
removed. Added an e2e spec (:action-plan prompt opens it, Escape closes);
suite now 28.

Verified: 51 unit + 28 e2e green, tsc + build clean, no behavior change.
